requirements.txt
streamlit
PyPDF2
transformers
torch
numpy
faiss-cpu

